requirements.txt
# Core PyTorch with MPS support for Apple Silicon
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Data processing and scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Machine learning utilities
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress bars and utilities
tqdm>=4.62.0

# Configuration management
PyYAML>=6.0

# Experiment tracking (optional)
wandb>=0.12.0

# Development tools (optional)
jupyter>=1.0.0
ipykernel>=6.0.0

# Testing and code quality (optional)
pytest>=6.0.0
black>=21.0.0
flake8>=3.8.0
mypy>=0.800.0
